requirements/requirements.txt
aiofiles==24.1.0
aiohttp==3.10.3
boto3==1.34.162
eqty==0.6.9
fastapi==0.111.0
gunicorn==22.0.0
hiredis==2.3.2
inflection==0.5.1
motor==3.4.0
openai==1.27.0
psutil==6.0.0
pymongo==4.7.3
pyOpenSSL==23.0.0
python-dotenv==1.0.1
pyyaml==6.0.1
redis==5.0.4
starlette==0.37.2
tenacity==8.3.0
uvicorn==0.29.0
websockets==12.0
ruamel.yaml==0.18.6
